+++ src/Time.hs
@@ -1,3 +1,7 @@
+-- SPDX-FileCopyrightText: 2019 Serokell <https://serokell.io>
+--
+-- SPDX-License-Identifier: MPL-2.0
+
 -- | This module reexports main functionality.
 --
 -- More information about @O'Clock@ features