+++ src/Time/Units.hs
@@ -66,7 +66,6 @@ module Time.Units
        , timeout
        ) where
 
-import Control.Applicative ((*>))
 import Control.Monad (unless)
 import Control.Monad.IO.Class (MonadIO, liftIO)
 import Data.Char (isDigit, isLetter)
@@ -81,18 +80,6 @@ import GHC.Real (denominator, numerator, (%))
 import GHC.TypeLits (KnownSymbol, Symbol, symbolVal)
 import Text.ParserCombinators.ReadP (ReadP, char, munch1, option, pfail, (+++))
 import Text.ParserCombinators.ReadPrec (ReadPrec, lift)
-
-#ifdef HAS_hashable
-import Data.Hashable (Hashable)
-#endif
-
-#ifdef HAS_deepseq
-import Control.DeepSeq (NFData)
-#endif
-
-#ifdef HAS_serialise
-import Codec.Serialise (Serialise (..))
-#endif
 
 #ifdef HAS_aeson
 import Data.Aeson (ToJSON (..), FromJSON (..), withText)
@@ -146,18 +133,6 @@ instance Monoid (Time (rat :: Rat)) where
     {-# INLINE mappend #-}
     mconcat = foldl' (<>) mempty
     {-# INLINE mconcat #-}
-
-#ifdef HAS_hashable
-instance Hashable (Time (rat :: Rat))
-#endif
-
-#ifdef HAS_deepseq
-instance NFData (Time (rat :: Rat))
-#endif
-
-#ifdef HAS_serialise
-instance Serialise (Time (rat :: Rat))
-#endif
 
 #ifdef HAS_aeson
 instance (KnownUnitName unit) => ToJSON (Time (unit :: Rat)) where