+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+<!--
+SPDX-FileCopyrightText: 2019 Serokell <https://serokell.io>
+
+SPDX-License-Identifier: MPL-2.0
+-->
+
 Change log
 ==========
 o'clock uses [PVP Versioning][1].